do the operations in models file
