+++ src/modules/User/model.ts
@@ -1,4 +1,8 @@
+import { PrismaClient } from '@prisma/client';
 import { objectType, inputObjectType } from 'nexus/dist';
+import { sign } from 'jsonwebtoken';
+import { compare, hash } from 'bcrypt';
+import { AUTH_SIGNUP_ENABLED, JWT_EXPIRATION_PERIOD, JWT_SECRET } from '../../env';
 
 export const User = objectType({
   name: 'User',
@@ -24,3 +28,45 @@ export const AuthPayload = objectType({
     t.field('user', { type: 'User' });
   },
 });
+
+export class UserOperations {
+  constructor(private readonly User: PrismaClient['user']) {}
+
+  async signIn(userNamePass: {username: string; password: string}) {
+    const { username, password } = userNamePass;
+    const user = await this.User.findUnique({
+      where: {
+        username,
+      },
+    });
+    if (!user) {
+      throw new Error('user not found');
+    }
+    const passwordValid = (await compare(password, user.password || '')) || false;
+    if (!passwordValid) {
+      throw new Error('invalid password');
+    }
+    return {
+      token: sign({ userId: user.id }, JWT_SECRET, { expiresIn: JWT_EXPIRATION_PERIOD }),
+      user,
+    };
+  }
+
+  async signUp(user: {username: string; password: string}) {
+    if (!AUTH_SIGNUP_ENABLED) {
+      throw new Error('Sign up is disabled');
+    }
+    const { username, password } = user;
+    const hashedPassword = await hash(password, 10);
+    const created = await this.User.create({
+      data: {
+        username,
+        password: hashedPassword,
+      },
+    });
+    return {
+      token: sign({ userId: created.id }, JWT_SECRET),
+      user: created,
+    };
+  }
+}
