+++ src/modules/User/resolvers.ts
@@ -1,7 +1,5 @@
 import { queryField, mutationField, nonNull } from 'nexus/dist';
-import { compare, hash } from 'bcrypt';
-import { sign } from 'jsonwebtoken';
-import { AUTH_SIGNUP_ENABLED, JWT_EXPIRATION_PERIOD, JWT_SECRET } from '../../env';
+import { UserOperations } from './model';
 
 export const me = queryField('me', {
   type: 'User',
@@ -20,23 +18,8 @@ export const signIn = mutationField('signIn', {
     user: nonNull('UserNamePass'),
   },
   resolve: async (_parent, { user: userNamePass }, ctx) => {
-    const { username, password } = userNamePass;
-    const user = await ctx.prisma.user.findUnique({
-      where: {
-        username,
-      },
-    });
-    if (!user) {
-      throw new Error('user not found');
-    }
-    const passwordValid = (await compare(password, user.password || '')) || false;
-    if (!passwordValid) {
-      throw new Error('invalid password');
-    }
-    return {
-      token: sign({ userId: user.id }, JWT_SECRET, { expiresIn: JWT_EXPIRATION_PERIOD }),
-      user,
-    };
+    const operations = new UserOperations(ctx.prisma.user);
+    return operations.signIn(userNamePass);
   },
 });
 
@@ -46,22 +29,7 @@ export const signUp = mutationField('signUp', {
     user: nonNull('UserNamePass'),
   },
   resolve: async (_parent, { user }, ctx) => {
-    if (!AUTH_SIGNUP_ENABLED) {
-      throw new Error('Sign up is disabled');
-    }
-    const { username, password } = user;
-    const hashedPassword = await hash(password, 10);
-
-    const created = await ctx.prisma.user.create({
-      data: {
-        username,
-        password: hashedPassword,
-      },
-    });
-
-    return {
-      token: sign({ userId: created.id }, JWT_SECRET),
-      user: created,
-    };
+    const operations = new UserOperations(ctx.prisma.user);
+    return operations.signUp(user);
   },
 });
